[SPARK-21867][CORE] Support async spilling in UnsafeShuffleWriter
Add a multi-shuffle sorter which supports asynchronous spilling during a shuffle external sort.
The benefit is that we can insert and sort/spill in parallel, reducing the overall latency for jobs that
are heavy on shuffling (as are many ads jobs). The multi-shuffle sorter is added between the UnsafeShuffleWriter
and ShuffleExternalSorter such that few changes are needed outside of this component, and as such, we can
see clearly there is little room for regressing other code paths.

The multi-shuffle sorter is enabled via a configuration flag, spark.shuffle.async.num.sorter (default 1)
If the value is 1, then the multi-shuffle sorter is not used, it must be configured to have multiple
sorters (>=2)

There is a design spec here attached to the jira.